A small vignette game exploring the many intricacies of love. Experience how a relationship through struggle and pain grows into fruition and into something meaningful and beautiful, culminating in a final 'Yes, I do'.
The game features many layers of depth for you to explore. Or, as one reviewer has put it:
"It's very artsy. :P Multiple layers and stuff."

And another:
"I can really see the emotion in this work."
"I cried on level 17."

This is try before you die software. Please attribute it as such.