Fix for node type registrations not working due to static library.
Through these changes the somewhat long standing issue of node type registrations not working due to vfx graph being compile as a static library.

The backstory is, the vfxNode*** object files were not being linked in, as the dependency traversal used by many C++ linkers only look at the object level whether a dependency exists and the object file should be linked in. If it isn't linked in, this means any global initializers in that unit are skipped. As we use global C++ objects which need to run their ctors during the initializers phase, this broke the automatic registration scheme, as nothing from the outside referenced the object files with node types directly.

This is now fixed by compiling all nodes in one big source file, and adding a function in this source file which gets called in constructVfxGraph. This call creates a dependency and makes sure the object file gets linked in, making sure ALL initializers run.